+++ source/vslib/components/inc/sinLookupTable.h
@@ -28,7 +28,7 @@ namespace vslib
             : Component("SinLookupTable", name, parent),
               m_function(
                   "data", this,
-                  fgc4::utils::generateFunction<double, double>(sin, 0.0, 2.0 * std::numbers::pi, number_points), true
+                  fgc4::utils::generateFunction<float, float>(sin, 0.0, 2.0 * std::numbers::pi, number_points), true
               )
         {
             assert(number_points >= 2);
@@ -38,7 +38,7 @@ namespace vslib
         //!
         //! @param input_x Value to be looked up in the table
         //! @return Interpolated function value closest to the input_x
-        [[nodiscard]] auto interpolate(double input_x)
+        [[nodiscard]] auto interpolate(float input_x)
         {
             return m_function.interpolate(input_x);
         }
@@ -47,13 +47,13 @@ namespace vslib
         //!
         //! @param input_x Value to be looked up in the table
         //! @return Interpolated function value closest to the input_x
-        [[nodiscard]] auto operator()(double input_x)
+        [[nodiscard]] auto operator()(float input_x)
         {
             return m_function.interpolate(input_x);
         }
 
       private:
         //!< Component providing the sine function storage and interpolation functionalities
-        PeriodicLookupTable<double, double> m_function;
+        PeriodicLookupTable<float, float> m_function;
     };
 }   // namespace vslib
